requirements.txt
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
beartype==0.22.2
cattrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.3
dagger-io==0.19.2
distro==1.9.0
docstring-parser==0.17.0
exceptiongroup==1.3.0
googleapis-common-protos==1.70.0
gql==4.0.0
graphql-core==3.2.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.79
langfuse==3.6.2
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.34
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-logging==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==5.29.5
pydantic==2.12.0
pydantic-core==2.41.1
pygments==2.19.2
pyyaml==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
sniffio==1.3.1
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
python-dotenv==1.0.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-mock==3.15.1
jinja2==3.1.6
google-generativeai==0.8.5
pillow==12.0.0
